+++ src/beatmap/parser.h
@@ -216,7 +216,7 @@ static int process_input(P*);
 	static int process_color(P*);
 		static int process_color_combo(P*);
 		static int parse_color(P*, struct oshu_color**);
-		static int parse_color_channel(P*, int*);
+		static int parse_color_channel(P*, double*);
 	static int process_hit_object(P*);
 		static int parse_hit_object(P*, struct oshu_hit**);
 			static int parse_common_hit(P*, struct oshu_hit*);
